fix: admin-filter-panel
have to test on staging because it requires a db
