+++ package.json
@@ -34,7 +34,6 @@
     "datatables.net": "^2.2.1",
     "datatables.net-dt": "^2.2.1",
     "datatables.net-react": "^1.0.0",
-    "date-fns": "^3.6.0",
     "dotenv": "^16.5.0",
     "duck-duck-scrape": "^2.2.6",
     "express": "^4.21.1",
@@ -55,7 +54,6 @@
     "prismjs": "^1.29.0",
     "rate-limiter-flexible": "^7.1.1",
     "react": "^18.3.1",
-    "react-date-range": "^2.0.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.1.5",
     "react-scripts": "^5.0.1",
